Handling icons for monitoring tools
